backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pymongo==4.15.3
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
httpx==0.27.2
openai==1.54.3
google-generativeai==0.8.3
elevenlabs==2.21.0
moviepy==1.0.3
ffmpeg-python==0.2.0
pillow==11.0.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
pydub==0.25.1
motor==3.7.1
python-slugify==8.0.4
watchfiles==1.1.1
websockets==15.0.1
